initial support for array formula import
However we still need to decide how we want to handle the cells in the
range of the array formula. ODF and OOXML write only the top left cell
into the document as a formula cell, all other cells are written as
normal value cells.

We have two ways to handle this:
* Leave this problem to the implementer of the interface
* Store internally all array ranges and decide there whether a value cell
  is a real value cell or the value represents a formula result.

The second approach breaks a bit the concept of orcus because we no
longer can easily import the content without knowing what has been
imported already. The first one however means that we move some of the
import problems to the implementer and the mentioned concepts are only
found in ODF and OOXML.