Have immediate parent context manage the life cycle of its own child contexts.
The top level context handler now only deletes the root xml context instance.
Each individual context class is now responsible for managing the life cycles
of its own child contexts that it returns.

With this, a context can re-use the same instance repeatedly as a child
context instead of creating each child context on the heap.