Add skeleton handler for pivot cache definition parts.
One thing to note is that the same pivot cache definition part is
referenced multiple times: once from the workbook.xml and twice or
more times from the pivotTable#.xml parts.  One way to handle this is
to have the OPC code to keep track of which parts have already been
read, and skip any attempts to read the same part twice.